ST_SA => MoX_SA
SA_Allocate_Space() => Allocate_Space()

+++ MoM_Rasm/MoX_MoM.C
@@ -1,8 +1,11 @@
 
 #include "MoX_MoM.H"
 
+#include <stdint.h>
 
-int RAM_Min_KB = RAM_MIN_KB;                    // MGC  dseg:A5C4  WZD  dseg:E5CE
+
+int16_t RAM_Min_KB = RAM_MIN_KB;                    // MGC  dseg:A5C4  WZD  dseg:E5CE
+
 
 char _init_game_message[] = "";
 char _leave_game_message[] = "Thank you for playing Master of Magic!";
